refactor: move K*s models out of ray logs.
[ORQSDK-852]
